mod/#155: mypage/main 패키징
